Migrate CduAngle from u16 to i16 (signed-fraction representation)
The CduAngle doc-comment claimed "u16 twos-complement count" and "full
revolution = 2^15 counts", but `u16` is unsigned and the conversion
arithmetic actually used 2^16 — the comment crystallised a half-finished
migration from the AGC's i15 fixed-point encoding to a native Rust type.

Adopt i16 consistently, mirroring the AGC's signed-fraction CDU encoding
the same way `f64` replaces the AGC's double-precision fixed point:

- CduAngle wraps an `i16`; full revolution = 2^16 counts (unchanged
  resolution, 1 LSB ≈ 0.0055°); range [-180°, +180°).
- Gimbal-lock detection uses the simpler `i16::wrapping_sub().unsigned_abs()`
  idiom (replaces the previous u16 min-of-two-arcs pattern).
- coarse_align_step drops three now-redundant `as i16` casts.
- Platform emulator `read_cdu` returns `[i16; 3]` directly (drops the
  intermediate `as i32 as i16 as u16` round-trip).
- Wire protocol `Msg::OpticsCdu` fields become `i16`. Bytes on the wire
  are unchanged (`to_le_bytes` produces identical bits), only the Rust
  type interpretation flips.
- BridgeState mirrors on both MCUs flip to `i16` accordingly.

Tests updated:
- Three CduAngle tests renamed and rewritten for signed-range semantics
  (e.g. tc_cdu_3_half_rev → tc_cdu_3_minus_half_rev; CduAngle(0x8000)
  → CduAngle(i16::MIN), expecting -π instead of +π — same physical
  angle modulo 2π).
- One new test tc_cdu_6_plus_max_count covers i16::MAX.
- Three out-of-range construction sites updated (65000 → -536,
  49152 → -16384, 32768 → i16::MIN), same physical angles preserved.

506 host tests pass; bare-metal builds clean in dev and release for both
the Nucleo-F767ZI board and the RP2040 bridge.

Resolves audit finding #6.

+++ agc-core/src/control/attitude.rs
@@ -59,8 +59,8 @@ impl AttitudeError {
 
 /// Estimate body angular rates in rad/s from two successive CDU readings.
 ///
-/// Uses two's-complement (wrapping) subtraction on the raw `u16` counts to
-/// handle the 0/65536 wrap-around correctly, then converts to rad/s.
+/// Uses two's-complement (wrapping) subtraction on the raw `i16` counts to
+/// handle the ±180° wrap-around correctly, then converts to rad/s.
 ///
 /// # Preconditions
 /// - `dt > 0.0`.  A zero or negative interval is a programming error; the
@@ -78,9 +78,9 @@ pub fn compute_body_rates(cdu_new: [CduAngle; 3], cdu_old: [CduAngle; 3], dt: f6
 
     let mut rates = [0.0_f64; 3];
     for i in 0..3 {
-        // Two's-complement subtraction: cast to i16 after wrapping_sub gives a
-        // signed delta in (-32768, +32767], correctly handling 0/TAU wrap-around.
-        let delta_counts = (cdu_new[i].0.wrapping_sub(cdu_old[i].0) as i16) as f64;
+        // i16::wrapping_sub returns a signed delta in [-32768, +32767],
+        // correctly handling the ±180° wrap-around.
+        let delta_counts = cdu_new[i].0.wrapping_sub(cdu_old[i].0) as f64;
         let delta_rad = delta_counts * (TAU / 65536.0);
         rates[i] = delta_rad / dt;
     }
@@ -341,7 +341,7 @@ mod tests {
     /// The PD torque must be negative on the roll axis (restoring) and zero elsewhere.
     #[test]
     fn tc_att_02_pure_roll_error() {
-        let five_deg_counts = (5.0_f64.to_radians() * 65536.0 / TAU) as u16;
+        let five_deg_counts = (5.0_f64.to_radians() * 65536.0 / TAU) as i16;
         let cdu = [CduAngle(five_deg_counts), CduAngle(0), CduAngle(0)];
 
         let error = compute_attitude_error(cdu, linalg::IDENTITY, linalg::IDENTITY);
@@ -422,8 +422,8 @@ mod tests {
 
         // Round-trip via compute_body_rates
         let dt = 0.1_f64;
-        let delta: u16 = ((omega_roll * dt) * 65536.0 / TAU).round() as u16;
-        let cdu_old = [CduAngle(0u16), CduAngle(0), CduAngle(0)];
+        let delta: i16 = ((omega_roll * dt) * 65536.0 / TAU).round() as i16;
+        let cdu_old = [CduAngle(0), CduAngle(0), CduAngle(0)];
         let cdu_new = [CduAngle(delta), CduAngle(0), CduAngle(0)];
         let estimated = compute_body_rates(cdu_new, cdu_old, dt);
 
@@ -524,7 +524,7 @@ mod tests {
     #[test]
     fn tc_att_sign_ci10_roll_sign_convention() {
         // Encode +1° as CDU counts for the outer (roll) gimbal
-        let one_deg_counts = (1.0_f64.to_radians() * 65536.0 / TAU).round() as u16;
+        let one_deg_counts = (1.0_f64.to_radians() * 65536.0 / TAU).round() as i16;
         let cdu = [CduAngle(one_deg_counts), CduAngle(0), CduAngle(0)];
 
         let error = compute_attitude_error(cdu, linalg::IDENTITY, linalg::IDENTITY);
